Clean up and reorganize
